requirements.txt
dj-database-url==0.4.2
Django==1.11.4
django-tastypie==0.14.0
gunicorn==19.7.1
whitenoise==3.3.1
